added getComputerChoice and playRockPaperScissors functions
getComputerChoice gets a random integer, 0,1,2 and based on
that integer, chooses rock, paper, or scissors.

playRockPaperScissors has two inputs, playerSelection and
computerSelection. playerSelection is a string that the user
types to select whether they'd like to chose rock, paper, or
scissors. computerSelection is also a string that uses
getComputerChoice() to randomise a computer choice.